hello_world/requirements.txt
Django==2.0.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.4
